AddRecord renamed, centered add form labels vertically.
